% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmR-package.R
\docType{data}
\name{adducts}
\alias{adducts}
\title{Adduct ion table}
\format{A tibble with 4 variables and 47 rows:
\describe{
   \item{adduct}{adduct type}
   \item{adduct}{adduct charge}
   \item{mult}{multiplier to apply when calculating mass}
   \item{mass}{mass difference to apply when calculating mass}
}}
\usage{
adducts
}
\description{
A table containing common adduct ion types and their associated mass differences
}
\keyword{datasets}
