% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmR-package.R
\name{alt_molform}
\alias{alt_molform}
\title{Get alternate molecular formulas to account for possible errors in
molecule standardization.}
\usage{
alt_molform(mol_form)
}
\arguments{
\item{mol_form}{A string containing the molecular formula}
}
\value{
a vector of alternate molecular formulas
}
\description{
Get alternate molecular formulas to account for possible errors in
molecule standardization.
}
