% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm-parse.R
\name{cfm_parse_spec.raw}
\alias{cfm_parse_spec.raw}
\title{Parse CFM precomputed spectra stored as raw vectors/binary blobs}
\usage{
\method{cfm_parse_spec}{raw}(spec, return_annotation = T)
}
\arguments{
\item{spec}{Result from CFM precompute}

\item{return_annotation}{T/F should annotation be returned}
}
\description{
Parse CFM precomputed spectra stored as raw vectors/binary blobs
}
\section{Functions}{
\itemize{
\item \code{cfm_parse_spec.raw}: Parse CFM prec-compute data stored as binary blob
}}

