% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm-predict.R
\name{cfm_parse_spec}
\alias{cfm_parse_spec}
\title{Parse CFM precomputed spectra to R objects}
\usage{
cfm_parse_spec(spec, return_annotation = T)
}
\arguments{
\item{spec}{Result from CFM precompute}

\item{return_annotation}{T/F should annotation be returned}
}
\value{
tibble with variables mz, intensity and data
}
\description{
Parse CFM precomputed spectra to R objects
}
