% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm-predict.R
\name{cfm_predict}
\alias{cfm_predict}
\title{CFM-predict}
\usage{
cfm_predict(
  input_smiles_or_inchi,
  prob_thresh_for_prune = 0.001,
  param_filename,
  config_filename,
  include_annotations = 0,
  output_filename = NULL,
  apply_post_processing = 1
)
}
\arguments{
\item{input_smiles_or_inchi}{The smiles or inchi string of the structure
whose spectra you want to predict.}

\item{prob_thresh_for_prune}{The probability below which to prune unlikely
fragmentations (default 0.001)}

\item{param_filename}{The filename where the parameters of a trained cfm
model can be found (if not given, assumes param_output.log in current
directory)}

\item{config_filename}{The filename where the configuration parameters of the
cfm model can be found (if not given, assumes param_config.txt in current
directory)}

\item{include_annotations}{Whether to include fragment information in the
output spectra (0 = NO (default), 1 = YES ). Note: ignored for msp/mgf
output.}

\item{output_filename}{The filename of the output spectra file to write to
(if not given, returns list from stdout).}

\item{apply_post_processing}{Whether or not to post-process predicted spectra
to take the top 80% of energy (at least 5 peaks), or the highest 30 peaks
(whichever comes first) (0 = OFF, 1 = ON (default) ).}
}
\value{
When a output filename is provided, the path to the file. If no
  filename is provided, results are extracted from the standard output into a
  named list.
}
\description{
CFM-predict
}
