% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm-read.R
\name{cfm_predict_readfile}
\alias{cfm_predict_readfile}
\title{Read cfm_predict results from output_filename}
\usage{
cfm_predict_readfile(output_filename)
}
\arguments{
\item{output_filename}{A text file containing the output from cfm_predict.}
}
\value{
List containing predicted spectra (and annotations, if present)
}
\description{
Read cfm_predict results from output_filename
}
