% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm-predict.R
\name{cfm_predict_stdout}
\alias{cfm_predict_stdout}
\title{Read cfm-predict standard output into R object}
\usage{
cfm_predict_stdout(rst)
}
\arguments{
\item{rst}{Standard out from a completed system2 call to cfm-predict}
}
\value{
List containing predicted spectra (and annotations, if present)
}
\description{
Read cfm-predict standard output into R object
}
