% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm-read.R
\name{cfm_read_batch}
\alias{cfm_read_batch}
\title{Read results of batch CFM prediction}
\usage{
cfm_read_batch(out_dir = NULL, table_name = NULL, output_path = NULL)
}
\arguments{
\item{out_dir}{A path to a directory containing text files generated by
\code{cfm_predict_batch}}

\item{table_name}{A name to be used for the resultant database table}

\item{output_path}{A path to a SQLite database file for writing the results}
}
\value{
Writes table to sql lite database file designated by output_path
}
\description{
Read results of batch CFM prediction
}
