% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm-read.R
\name{cfm_read_db}
\alias{cfm_read_db}
\title{Read CFM precomputed spectra from a database created by cfm_read_batch}
\usage{
cfm_read_db(
  db_file = NULL,
  table_name = NULL,
  ID = NULL,
  return_annotation = F
)
}
\arguments{
\item{db_file}{[string] Full path to the database file}

\item{table_name}{[string] Name of table containing the CFM spectra to read}

\item{ID}{[string] A vector of ID(s) to retrieve from the database}

\item{return_annotation}{[logical] Should spectra be returned with annotations}
}
\value{
A named list of spectra
}
\description{
Read CFM precomputed spectra from a database created by cfm_read_batch
}
