% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm-search.R
\name{cfm_search}
\alias{cfm_search}
\title{Spectral similarity searching against a pre-computed CFM library}
\usage{
cfm_search(
  query_spectrum,
  db_file = NULL,
  table_name = NULL,
  ID = NULL,
  fun = c("common", "cor", "dotproduct"),
  bin = 1
)
}
\arguments{
\item{query_spectrum}{An MSnbase::Spectrum2 object to search against the
library}

\item{db_file}{A database file containing the CFM precomputed spectra}

\item{table_name}{The name of the table containing the precomputed spectra}

\item{ID}{Values to filter by. Corresponds to the ID column in the table. If
NULL than all spectra from the library are compared.}

\item{fun}{Function to use for spectral comparrison (common, cor, or
dotproduct). See MSnbase::compareSpectra.}

\item{bin}{Bin size for peak patching. See MSnbase::compareSpectra for
details.}
}
\value{
A tibble with variables ID, %fun_similarity
}
\description{
Spectral similarity searching against a pre-computed CFM library
}
