% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm-to-mzvault.R
\name{cfm_to_mzvault}
\alias{cfm_to_mzvault}
\title{Create a spectral library in mzVault format from CFM predicted spectra}
\usage{
cfm_to_mzvault(
  input_db,
  mol_tab,
  mol_idx,
  identifier_tab,
  cfmpred_pos,
  cfmpred_neg,
  ppm = 5
)
}
\arguments{
\item{input_db}{A full path to a database file containing predicted spectra
and molecules}

\item{mol_tab}{The name of the table containing molecule information.}

\item{mol_idx}{The name of the variable used for table indexing. Should match
between spectra and molecule tables.}

\item{identifier_tab}{(optional) The name of the table containing identifiers
to append to the library records. Must contain an InChI Key column.}

\item{cfmpred_pos}{The name of the table containing the CMF-ID positive ion
predicted spectra.}

\item{cfmpred_neg}{The name of the table containing the CFM-ID negative ion
predicted spectra.}

\item{ppm}{Part-per-million mass error to use for fragment ion matching.}
}
\value{

}
\description{
Create a spectral library in mzVault format from CFM predicted spectra
}
