% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmR-package.R
\name{clean_form}
\alias{clean_form}
\title{Clean molecular formula}
\usage{
clean_form(molform, withspace = FALSE, as_cts = FALSE)
}
\arguments{
\item{molform}{molecular formula or a vector of molecular formulas.}

\item{withspace}{return the cleaned formula with or without spaces (T/F)}

\item{as_cts}{T/F return formula as a string or a named vector}
}
\value{
cleaned molecular formula(s)
}
\description{
Clean molecular formula
}
\examples{
clean_form("C15H17BrN3+", withspace = TRUE)
clean_form("C15H17BrN3+", as_cts = TRUE)
}
