% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm-parse.R
\name{cfm_parse_spec}
\alias{cfm_parse_spec}
\title{Parse CFM precomputed spectra to R objects}
\usage{
cfm_parse_spec(spec, return_annotation = T)
}
\arguments{
\item{spec}{Result from CFM precompute}

\item{return_annotation}{T/F should annotation be returned}
}
\value{
the predicted mass spectrum
}
\description{
Returns a Spectrum2 object if return_annotation = F and a tibble with
columns (mz,intensity,data) if return_annotation = T. The data column
contains the predicted intensity at three energies and the SMILES string for
the peak annotation.
}
