% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm-search.R
\name{cfm_search}
\alias{cfm_search}
\title{Spectral similarity searching against a pre-computed CFM library}
\usage{
cfm_search(
  query_spectrum,
  mol_form = NULL,
  db_file = NULL,
  spec_table_name = NULL,
  mol_table_name = NULL,
  molprop_table_name = NULL,
  ID = NULL,
  fun = c("common", "cor", "dotproduct"),
  bin = 1,
  pol = c("negative", "positive"),
  ppm = 10,
  excl_precur = F
)
}
\arguments{
\item{query_spectrum}{An MSnbase::Spectrum2 object to search against the
library}

\item{mol_form}{The molecular formula assigned to the query spectrum.}

\item{db_file}{A database file containing the CFM precomputed spectra}

\item{spec_table_name}{The name of the table containing the precomputed spectra}

\item{mol_table_name}{The name of the table containing molecules}

\item{molprop_table_name}{The name of the table continaing molecular properties}

\item{ID}{Values to filter by. Corresponds to the ID column in the table. If
NULL than all spectra from the library are compared.}

\item{fun}{Function to use for spectral comparrison (common, cor, or
dotproduct). See MSnbase::compareSpectra.}

\item{bin}{Bin size for peak patching. See MSnbase::compareSpectra for
details.}

\item{pol}{Polarity (positive or negative)}

\item{ppm}{Mass error for filtering molecules by neutral mass}

\item{excl_precur}{Should precursor ion from query spectrum be excluded. (T/F)}
}
\value{
A tibble with variables ID, %fun_similarity
}
\description{
Spectral similarity searching against a pre-computed CFM library
}
